+++ addons/recoils/CfgWeapons.hpp
@@ -4,37 +4,11 @@ class CfgWeapons {
 
     // 5.45
     #include "bohemia\545\ak.hpp"
-    #include "cup\545\ak.hpp"
-    #include "cup\545\ak12.hpp"
 
     // 5.56
     #include "bohemia\556\f2000.hpp"
     #include "bohemia\556\hk416.hpp"
     #include "bohemia\556\tar21.hpp"
-    #include "cup\556\acr.hpp"
-    #include "cup\556\ak.hpp"
-    #include "cup\556\ak19.hpp"
-    #include "cup\556\aug.hpp"
-    #include "cup\556\cz.hpp"
-    #include "cup\556\famas.hpp"
-    #include "cup\556\fort.hpp"
-    #include "cup\556\g36.hpp"
-    #include "cup\556\galil.hpp"
-    #include "cup\556\hk416.hpp"
-    #include "cup\556\l85.hpp"
-    #include "cup\556\m4.hpp"
-    #include "cup\556\m16.hpp"
-    #include "cup\556\m27.hpp"
-    #include "cup\556\mg36.hpp"
-    #include "cup\556\mk12.hpp"
-    #include "cup\556\mk18.hpp"
-    #include "cup\556\scar.hpp"
-    #include "niarms\556\acr.hpp"
-    #include "niarms\556\aug.hpp"
-    #include "niarms\556\m4.hpp"
-    #include "niarms\556\sig.hpp"
-    #include "ws\556\velko.hpp"
-    #include "ws\556\xms.hpp"
 
     // 6.5
     #include "bohemia\65\katiba.hpp"
@@ -43,53 +17,14 @@ class CfgWeapons {
     #include "bohemia\65\qbu.hpp"
     #include "bohemia\65\type115.hpp"
 
-    // 6.8
-    #include "cup\68\acr.hpp"
-
     // 7.62x39
     #include "bohemia\762R\ak.hpp"
-    #include "cup\762R\ak.hpp"
-    #include "cup\762R\ak15.hpp"
-    #include "cup\762R\cz.hpp"
-    #include "cup\762R\groza.hpp"
-    #include "cup\762R\rpk.hpp"
-    #include "niarms\762R\sig.hpp"
-    #include "pi\762R\rk.hpp"
-    #include "ws\762R\galat.hpp"
 
     // 7.62x51
     #include "bohemia\762N\ebr.hpp"
     #include "bohemia\762N\hk417.hpp"
     #include "bohemia\762N\m14.hpp"
     #include "bohemia\762N\sig.hpp"
-    #include "cup\762N\fal.hpp"
-    #include "cup\762N\g3.hpp"
-    #include "cup\762N\galil.hpp"
-    #include "cup\762N\gewehr.hpp"
-    #include "cup\762N\hk417.hpp"
-    #include "cup\762N\l129a1.hpp"
-    #include "cup\762N\m14.hpp"
-    #include "cup\762N\m110.hpp"
-    #include "cup\762N\romat.hpp"
-    #include "cup\762N\sa58.hpp"
-    #include "cup\762N\scar.hpp"
-    #include "cup\762N\stg.hpp"
-    #include "ws\762N\slr.hpp"
-
-    // 7.62x54r
-    #include "cup\762_54\svd.hpp"
-
-    // 12G Shotguns
-    #include "cup\12G\m1014.hpp"
-    #include "cup\12G\saiga12.hpp"
-    #include "ws\12G\aa40.hpp"
-
-    // .300 Blackout
-    #include "niarms\blackout\ar15.hpp"
-    #include "niarms\blackout\honeybadger.hpp"
-
-    // .300 WM
-    #include "pi\300WM\falkor.hpp"
 
     // .338
     #include "bohemia\338\mar10.hpp"
