+++ addons/recoils/baseclasses.hpp
@@ -1,4 +1,3 @@
-// BI
 // AKs
 class arifle_AKS_base_F;
 class arifle_AKM_base_F;
@@ -64,124 +63,9 @@ class Tavor_base_F;
 // Type 115
 class arifle_ARX_base_F;
 
-// CUP
-// ACRs
-class CUP_arifle_ACR_Short_BASE_556;
-class CUP_arifle_ACR_EGLM_SHORT_BASE_556;
-class CUP_arifle_ACR_BASE_556;
-class CUP_arifle_ACR_DMR_BASE_556;
-class CUP_arifle_ACR_EGLM_BASE_556;
-class CUP_arifle_ACR_Short_BASE_68;
-class CUP_arifle_ACR_EGLM_SHORT_BASE_68;
-class CUP_arifle_ACR_BASE_68;
-class CUP_arifle_ACR_EGLM_BASE_68;
-class CUP_arifle_ACR_DMR_BASE_68;
-
-// AKs
-class CUP_arifle_AK_Base;
-class CUP_arifle_AKS_Base;
-class CUP_arifle_AK107_Base;
-
-// AUG
-class CUP_arifle_AUG_Base;
-
-// CZs
-class CUP_arifle_CZ805_Base;
-class CUP_CZ_BREN2_Base;
-class CUP_CZ_BREN2_GL_Base;
-class CUP_CZ_BREN2_762_Base;
-class CUP_CZ_BREN2_762_GL_Base;
-
-// FAMAS
-class CUP_Famas_Base;
-
-// Fort / X95
+// Fort
 class arifle_TRG20_F;
 class arifle_TRG21_F;
-class CUP_arifle_X95;
-class CUP_arifle_X95_Grippod;
-
-// G3
-class CUP_arifle_G3_Base;
-
-// G36
-class CUP_arifle_G36_Base_CarryhandleOptics;
-class CUP_arifle_G36_Base;
-
-// Galil
-class CUP_arifle_GALIL_BASE;
-
-// Groza
-class CUP_arifle_OTS14_GROZA_762_base;
-
-// HK416
-class CUP_arifle_HK416_145_Base;
-class CUP_arifle_HK416_11_Base;
-
-// HK417
-class CUP_arifle_HK417_Base;
-
-// L85
-class CUP_l85a2;
-class CUP_l85a2_ris_ng;
-class CUP_l85a2_ris;
-class CUP_l85a2_ugl;
-
-// M16
-class CUP_arifle_XM16E1;
-class CUP_arifle_M4_Base;
-class CUP_arifle_M16_Base;
-
-// M4
-class CUP_arifle_M4A1_BUIS_Base;
-
-// SCAR
-class CUP_arifle_SCAR_L_Base;
-class CUP_arifle_Mk17_Base;
-
-// M1014
-class CUP_sgun_M1014_base;
 
 // Saiga
 class Rifle_Base_F;
-
-// NiArms
-// ACR-E
-class hlc_acr556_base;
-class hlc_rifle_ACR_SBR_black;
-class hlc_rifle_ACR68_SBR_black;
-class hlc_acr_base;
-
-// AUG
-class hlc_aug_base;
-class hlc_rifle_auga3;
-
-// Honey Badger
-class hlc_rifle_honeybase;
-
-// M4 Variants
-class hlc_ar15_base;
-class hlc_rifle_CQBR;
-
-// SG55X
-class hlc_sg550_base;
-
-// Project Infinite
-// Falkor
-class bnae_falkor_base;
-
-// RK95s
-class bnae_rk95_base;
-
-// Western Sahara
-// AA40
-class sgun_aa40_base_lxWS;
-
-// Galat
-class arifle_Galat_base_lxWS;
-
-// Velko
-class arifle_Velko_base_lxWS;
-
-// SLR
-class arifle_SLR_V_lxWS;
