+++ addons/recoils/niarms/script_component.hpp
@@ -1,4 +1,4 @@
-#define COMPONENT recoils
-#define COMPONENT_BEAUTIFIED Recoils
+#define COMPONENT recoils_hlc
+#define COMPONENT_BEAUTIFIED Recoils NIArms
 #include "\x\tacgt\addons\main\script_mod.hpp"
 #include "\x\tacgt\addons\main\script_macros.hpp"
