+++ addons/recoils/pi/config.cpp
@@ -8,8 +8,12 @@ class CfgPatches {
         magazines[] = {};
         requiredVersion = REQUIRED_VERSION;
         requiredAddons[] = {
-            "tacgt_main"
+            "tacgt_main",
+            "tacgt_recoils",
+            "bnae_falkor",
+            "bnae_RK95"
         };
+        skipWhenMissingDependencies = 1;
         author = ECSTRING(main,Authors);
         authors[] = {"Mike"};
         url = ECSTRING(main,URL);
@@ -17,5 +21,13 @@ class CfgPatches {
     };
 };
 
-#include "CfgRecoils.hpp"
-#include "CfgWeapons.hpp"
+class CfgWeapons {
+    // Falkor
+    class bnae_falkor_base;
+
+    // RK95s
+    class bnae_rk95_base;
+
+    #include "762R\rk.hpp"
+    #include "300WM\falkor.hpp"
+};
