+++ addons/recoils/pi/script_component.hpp
@@ -1,4 +1,4 @@
-#define COMPONENT recoils
-#define COMPONENT_BEAUTIFIED Recoils
+#define COMPONENT recoils_pi
+#define COMPONENT_BEAUTIFIED Recoils Project Infinite
 #include "\x\tacgt\addons\main\script_mod.hpp"
 #include "\x\tacgt\addons\main\script_macros.hpp"
