+++ addons/recoils/ws/script_component.hpp
@@ -1,4 +1,4 @@
-#define COMPONENT recoils
-#define COMPONENT_BEAUTIFIED Recoils
+#define COMPONENT recoils_ws
+#define COMPONENT_BEAUTIFIED Recoils Western Sahara
 #include "\x\tacgt\addons\main\script_mod.hpp"
 #include "\x\tacgt\addons\main\script_macros.hpp"
